Revert "Improve codecov (#1384)"
This reverts commit 060a9e5726e35db53885e9f04b88a3c77ff51845.

+++ service/routing_config_v2.go
@@ -19,16 +19,12 @@ package service
 
 import (
 	"context"
-	"fmt"
 	"strconv"
-	"time"
 
-	"github.com/golang/protobuf/jsonpb"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/any"
 	"github.com/golang/protobuf/ptypes/wrappers"
 	apimodel "github.com/polarismesh/specification/source/go/api/v1/model"
-	"github.com/polarismesh/specification/source/go/api/v1/security"
 	apiservice "github.com/polarismesh/specification/source/go/api/v1/service_manage"
 	apitraffic "github.com/polarismesh/specification/source/go/api/v1/traffic_manage"
 	"go.uber.org/zap"
@@ -36,14 +32,36 @@ import (
 	cachetypes "github.com/polarismesh/polaris/cache/api"
 	apiv1 "github.com/polarismesh/polaris/common/api/v1"
 	"github.com/polarismesh/polaris/common/model"
-	authcommon "github.com/polarismesh/polaris/common/model/auth"
 	commonstore "github.com/polarismesh/polaris/common/store"
 	"github.com/polarismesh/polaris/common/utils"
+)
+
+var (
+	// RoutingConfigV2FilterAttrs router config filter attrs
+	RoutingConfigV2FilterAttrs = map[string]bool{
+		"id":                    true,
+		"name":                  true,
+		"service":               true,
+		"namespace":             true,
+		"source_service":        true,
+		"destination_service":   true,
+		"source_namespace":      true,
+		"destination_namespace": true,
+		"enable":                true,
+		"offset":                true,
+		"limit":                 true,
+		"order_field":           true,
+		"order_type":            true,
+	}
 )
 
 // CreateRoutingConfigsV2 Create a routing configuration
 func (s *Server) CreateRoutingConfigsV2(
 	ctx context.Context, req []*apitraffic.RouteRule) *apiservice.BatchWriteResponse {
+	if err := checkBatchRoutingConfigV2(req); err != nil {
+		return err
+	}
+
 	resp := apiv1.NewBatchWriteResponse(apimodel.Code_ExecuteSuccess)
 	for _, entry := range req {
 		apiv1.Collect(resp, s.createRoutingConfigV2(ctx, entry))
@@ -54,6 +72,10 @@ func (s *Server) CreateRoutingConfigsV2(
 
 // createRoutingConfigV2 Create a routing configuration
 func (s *Server) createRoutingConfigV2(ctx context.Context, req *apitraffic.RouteRule) *apiservice.Response {
+	if resp := checkRoutingConfigV2(req); resp != nil {
+		return resp
+	}
+
 	conf, err := Api2RoutingConfigV2(req)
 	if err != nil {
 		log.Error("[Routing][V2] parse routing config v2 from request for create",
@@ -67,11 +89,8 @@ func (s *Server) createRoutingConfigV2(ctx context.Context, req *apitraffic.Rout
 		return apiv1.NewResponse(commonstore.StoreCode2APICode(err))
 	}
 
-	s.RecordHistory(ctx, routeRuleRecordEntry(ctx, req, conf, model.OCreate))
-	_ = s.afterRuleResource(ctx, model.RRouting, authcommon.ResourceEntry{
-		ID:   req.GetId(),
-		Type: security.ResourceType_RouteRules,
-	}, false)
+	s.RecordHistory(ctx, routingV2RecordEntry(ctx, req, conf, model.OCreate))
+
 	req.Id = conf.ID
 	return apiv1.NewRouterResponse(apimodel.Code_ExecuteSuccess, req)
 }
@@ -79,6 +98,10 @@ func (s *Server) createRoutingConfigV2(ctx context.Context, req *apitraffic.Rout
 // DeleteRoutingConfigsV2 Batch delete routing configuration
 func (s *Server) DeleteRoutingConfigsV2(
 	ctx context.Context, req []*apitraffic.RouteRule) *apiservice.BatchWriteResponse {
+	if err := checkBatchRoutingConfigV2(req); err != nil {
+		return err
+	}
+
 	out := apiv1.NewBatchWriteResponse(apimodel.Code_ExecuteSuccess)
 	for _, entry := range req {
 		resp := s.deleteRoutingConfigV2(ctx, entry)
@@ -90,27 +113,38 @@ func (s *Server) DeleteRoutingConfigsV2(
 
 // DeleteRoutingConfigV2 Delete a routing configuration
 func (s *Server) deleteRoutingConfigV2(ctx context.Context, req *apitraffic.RouteRule) *apiservice.Response {
+	if resp := checkRoutingConfigIDV2(req); resp != nil {
+		return resp
+	}
+
+	// Determine whether the current routing rules are only converted from the memory transmission in the V1 version
+	if _, ok := s.Cache().RoutingConfig().IsConvertFromV1(req.Id); ok {
+		resp := s.transferV1toV2OnModify(ctx, req)
+		if resp.GetCode().GetValue() != uint32(apimodel.Code_ExecuteSuccess) {
+			return resp
+		}
+	}
+
 	if err := s.storage.DeleteRoutingConfigV2(req.Id); err != nil {
 		log.Error("[Routing][V2] delete routing config v2 store layer",
 			utils.RequestID(ctx), zap.Error(err))
 		return apiv1.NewResponse(commonstore.StoreCode2APICode(err))
 	}
 
-	s.RecordHistory(ctx, routeRuleRecordEntry(ctx, req, &model.RouterConfig{
+	s.RecordHistory(ctx, routingV2RecordEntry(ctx, req, &model.RouterConfig{
 		ID:   req.GetId(),
 		Name: req.GetName(),
 	}, model.ODelete))
-
-	_ = s.afterRuleResource(ctx, model.RRouting, authcommon.ResourceEntry{
-		ID:   req.GetId(),
-		Type: security.ResourceType_RouteRules,
-	}, true)
 	return apiv1.NewRouterResponse(apimodel.Code_ExecuteSuccess, req)
 }
 
 // UpdateRoutingConfigsV2 Batch update routing configuration
 func (s *Server) UpdateRoutingConfigsV2(
 	ctx context.Context, req []*apitraffic.RouteRule) *apiservice.BatchWriteResponse {
+	if err := checkBatchRoutingConfigV2(req); err != nil {
+		return err
+	}
+
 	out := apiv1.NewBatchWriteResponse(apimodel.Code_ExecuteSuccess)
 	for _, entry := range req {
 		resp := s.updateRoutingConfigV2(ctx, entry)
@@ -122,6 +156,21 @@ func (s *Server) UpdateRoutingConfigsV2(
 
 // updateRoutingConfigV2 Update a single routing configuration
 func (s *Server) updateRoutingConfigV2(ctx context.Context, req *apitraffic.RouteRule) *apiservice.Response {
+	// If V2 routing rules to be modified are from the V1 rule in the cache, need to do the following steps first
+	// step 1: Turn the V1 rule to the real V2 rule
+	// step 2: Find the corresponding route to the V2 rules to be modified in the V1 rules, set their rules ID
+	// step 3: Store persistence
+	if _, ok := s.Cache().RoutingConfig().IsConvertFromV1(req.Id); ok {
+		resp := s.transferV1toV2OnModify(ctx, req)
+		if resp.GetCode().GetValue() != uint32(apimodel.Code_ExecuteSuccess) {
+			return resp
+		}
+	}
+
+	if resp := checkUpdateRoutingConfigV2(req); resp != nil {
+		return resp
+	}
+
 	// Check whether the routing configuration exists
 	conf, err := s.storage.GetRoutingConfigV2WithID(req.Id)
 	if err != nil {
@@ -147,7 +196,7 @@ func (s *Server) updateRoutingConfigV2(ctx context.Context, req *apitraffic.Rout
 		return apiv1.NewResponse(commonstore.StoreCode2APICode(err))
 	}
 
-	s.RecordHistory(ctx, routeRuleRecordEntry(ctx, req, reqModel, model.OUpdate))
+	s.RecordHistory(ctx, routingV2RecordEntry(ctx, req, reqModel, model.OUpdate))
 	return apiv1.NewResponse(apimodel.Code_ExecuteSuccess)
 }
 
@@ -178,11 +227,6 @@ func (s *Server) QueryRoutingConfigsV2(ctx context.Context, query map[string]str
 	return resp
 }
 
-// GetAllRouterRules Query all router_rule rules
-func (s *Server) GetAllRouterRules(ctx context.Context) *apiservice.BatchQueryResponse {
-	return nil
-}
-
 // EnableRoutings batch enable routing rules
 func (s *Server) EnableRoutings(ctx context.Context, req []*apitraffic.RouteRule) *apiservice.BatchWriteResponse {
 	out := apiv1.NewBatchWriteResponse(apimodel.Code_ExecuteSuccess)
@@ -195,6 +239,17 @@ func (s *Server) EnableRoutings(ctx context.Context, req []*apitraffic.RouteRule
 }
 
 func (s *Server) enableRoutings(ctx context.Context, req *apitraffic.RouteRule) *apiservice.Response {
+	if resp := checkRoutingConfigIDV2(req); resp != nil {
+		return resp
+	}
+
+	if _, ok := s.Cache().RoutingConfig().IsConvertFromV1(req.Id); ok {
+		resp := s.transferV1toV2OnModify(ctx, req)
+		if resp.GetCode().GetValue() != uint32(apimodel.Code_ExecuteSuccess) {
+			return resp
+		}
+	}
+
 	conf, err := s.storage.GetRoutingConfigV2WithID(req.Id)
 	if err != nil {
 		log.Error("[Routing][V2] get routing config v2 store layer",
@@ -214,13 +269,83 @@ func (s *Server) enableRoutings(ctx context.Context, req *apitraffic.RouteRule) 
 		return apiv1.NewResponse(commonstore.StoreCode2APICode(err))
 	}
 
-	s.RecordHistory(ctx, routeRuleRecordEntry(ctx, req, conf, model.OUpdate))
+	s.RecordHistory(ctx, routingV2RecordEntry(ctx, req, conf, model.OUpdate))
+	return apiv1.NewResponse(apimodel.Code_ExecuteSuccess)
+}
+
+// transferV1toV2OnModify When enabled or prohibited for the V2 rules, the V1 rules need to be converted to V2 rules
+// and execute persistent storage
+func (s *Server) transferV1toV2OnModify(ctx context.Context, req *apitraffic.RouteRule) *apiservice.Response {
+	svcId, _ := s.Cache().RoutingConfig().IsConvertFromV1(req.Id)
+	v1conf, err := s.storage.GetRoutingConfigWithID(svcId)
+	if err != nil {
+		log.Error("[Routing][V2] get routing config v1 store layer",
+			utils.RequestID(ctx), zap.Error(err))
+		return apiv1.NewResponse(commonstore.StoreCode2APICode(err))
+	}
+	if v1conf != nil {
+		svc, err := s.loadServiceByID(svcId)
+		if svc == nil {
+			log.Error("[Routing][V2] convert routing config v1 to v2 find svc",
+				utils.RequestID(ctx), zap.Error(err))
+			return apiv1.NewResponse(apimodel.Code_NotFoundService)
+		}
+
+		inV2, outV2, err := model.ConvertRoutingV1ToExtendV2(svc.Name, svc.Namespace, v1conf)
+		if err != nil {
+			log.Error("[Routing][V2] convert routing config v1 to v2",
+				utils.RequestID(ctx), zap.Error(err))
+			return apiv1.NewResponse(apimodel.Code_ExecuteException)
+		}
+
+		formatApi := func(rules []*model.ExtendRouterConfig) ([]*apitraffic.RouteRule, *apiservice.Response) {
+			ret := make([]*apitraffic.RouteRule, 0, len(rules))
+			for i := range rules {
+				item, err := rules[i].ToApi()
+				if err != nil {
+					log.Error("[Routing][V2] convert routing config v1 to v2, format v2 to api",
+						utils.RequestID(ctx), zap.Error(err))
+					return nil, apiv1.NewResponse(apimodel.Code_ExecuteException)
+				}
+				ret = append(ret, item)
+			}
+
+			return ret, nil
+		}
+
+		inDatas, resp := formatApi(inV2)
+		if resp != nil {
+			return resp
+		}
+		outDatas, resp := formatApi(outV2)
+		if resp != nil {
+			return resp
+		}
+
+		if resp := s.saveRoutingV1toV2(ctx, svcId, inDatas, outDatas); resp.GetCode().GetValue() != apiv1.ExecuteSuccess {
+			return apiv1.NewResponse(apimodel.Code(resp.GetCode().GetValue()))
+		}
+	}
+
 	return apiv1.NewResponse(apimodel.Code_ExecuteSuccess)
 }
 
 // parseServiceArgs The query conditions of the analysis service
-func parseRoutingArgs(filter map[string]string, ctx context.Context) (*cachetypes.RoutingArgs, *apiservice.Response) {
-	offset, limit, _ := utils.ParseOffsetAndLimit(filter)
+func parseRoutingArgs(query map[string]string, ctx context.Context) (*cachetypes.RoutingArgs, *apiservice.Response) {
+	offset, limit, err := utils.ParseOffsetAndLimit(query)
+	if err != nil {
+		return nil, apiv1.NewResponse(apimodel.Code_InvalidParameter)
+	}
+
+	filter := make(map[string]string)
+	for key, value := range query {
+		if _, ok := RoutingConfigV2FilterAttrs[key]; !ok {
+			log.Errorf("[Routing][V2][Query] attribute(%s) is not allowed", key)
+			return nil, apiv1.NewResponse(apimodel.Code_InvalidParameter)
+		}
+		filter[key] = value
+	}
+
 	res := &cachetypes.RoutingArgs{
 		Filter:     filter,
 		Name:       filter["name"],
@@ -252,6 +377,120 @@ func parseRoutingArgs(filter map[string]string, ctx context.Context) (*cachetype
 	}
 	log.Infof("[Service][Routing][Query] routing query args: %+v", res)
 	return res, nil
+}
+
+// checkBatchRoutingConfig Check batch request
+func checkBatchRoutingConfigV2(req []*apitraffic.RouteRule) *apiservice.BatchWriteResponse {
+	if len(req) == 0 {
+		return apiv1.NewBatchWriteResponse(apimodel.Code_EmptyRequest)
+	}
+
+	if len(req) > MaxBatchSize {
+		return apiv1.NewBatchWriteResponse(apimodel.Code_BatchSizeOverLimit)
+	}
+
+	return nil
+}
+
+// checkRoutingConfig Check the validity of the basic parameter of the routing configuration
+func checkRoutingConfigV2(req *apitraffic.RouteRule) *apiservice.Response {
+	if req == nil {
+		return apiv1.NewRouterResponse(apimodel.Code_EmptyRequest, req)
+	}
+
+	if err := checkRoutingNameAndNamespace(req); err != nil {
+		return err
+	}
+
+	if err := checkRoutingConfigPriorityV2(req); err != nil {
+		return err
+	}
+
+	if err := checkRoutingPolicyV2(req); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// checkUpdateRoutingConfigV2 Check the validity of the basic parameter of the routing configuration
+func checkUpdateRoutingConfigV2(req *apitraffic.RouteRule) *apiservice.Response {
+	if resp := checkRoutingConfigIDV2(req); resp != nil {
+		return resp
+	}
+
+	if err := checkRoutingNameAndNamespace(req); err != nil {
+		return err
+	}
+
+	if err := checkRoutingConfigPriorityV2(req); err != nil {
+		return err
+	}
+
+	if err := checkRoutingPolicyV2(req); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func checkRoutingNameAndNamespace(req *apitraffic.RouteRule) *apiservice.Response {
+	if err := utils.CheckDbStrFieldLen(utils.NewStringValue(req.GetName()), MaxDbRoutingName); err != nil {
+		return apiv1.NewRouterResponse(apimodel.Code_InvalidRoutingName, req)
+	}
+
+	if err := utils.CheckDbStrFieldLen(utils.NewStringValue(req.GetNamespace()),
+		MaxDbServiceNamespaceLength); err != nil {
+		return apiv1.NewRouterResponse(apimodel.Code_InvalidNamespaceName, req)
+	}
+
+	return nil
+}
+
+func checkRoutingConfigIDV2(req *apitraffic.RouteRule) *apiservice.Response {
+	if req == nil {
+		return apiv1.NewRouterResponse(apimodel.Code_EmptyRequest, req)
+	}
+
+	if req.Id == "" {
+		return apiv1.NewResponse(apimodel.Code_InvalidRoutingID)
+	}
+
+	return nil
+}
+
+func checkRoutingConfigPriorityV2(req *apitraffic.RouteRule) *apiservice.Response {
+	if req == nil {
+		return apiv1.NewRouterResponse(apimodel.Code_EmptyRequest, req)
+	}
+
+	if req.Priority > 10 {
+		return apiv1.NewResponse(apimodel.Code_InvalidRoutingPriority)
+	}
+
+	return nil
+}
+
+func checkRoutingPolicyV2(req *apitraffic.RouteRule) *apiservice.Response {
+	if req == nil {
+		return apiv1.NewRouterResponse(apimodel.Code_EmptyRequest, req)
+	}
+
+	if req.GetRoutingPolicy() != apitraffic.RoutingPolicy_RulePolicy {
+		return apiv1.NewRouterResponse(apimodel.Code_InvalidRoutingPolicy, req)
+	}
+
+	// Automatically supplement @Type attribute according to Policy
+	if req.RoutingConfig.TypeUrl == "" {
+		if req.GetRoutingPolicy() == apitraffic.RoutingPolicy_RulePolicy {
+			req.RoutingConfig.TypeUrl = model.RuleRoutingTypeUrl
+		}
+		if req.GetRoutingPolicy() == apitraffic.RoutingPolicy_MetadataPolicy {
+			req.RoutingConfig.TypeUrl = model.MetaRoutingTypeUrl
+		}
+	}
+
+	return nil
 }
 
 // Api2RoutingConfigV2 Convert the API parameter to internal data structure
@@ -291,23 +530,4 @@ func marshalRoutingV2toAnySlice(routings []*model.ExtendRouterConfig) ([]*any.An
 	}
 
 	return ret, nil
-}
-
-// routeRuleRecordEntry Construction of RoutingConfig's record Entry
-func routeRuleRecordEntry(ctx context.Context, req *apitraffic.RouteRule, md *model.RouterConfig,
-	opt model.OperationType) *model.RecordEntry {
-
-	marshaler := jsonpb.Marshaler{}
-	detail, _ := marshaler.MarshalToString(req)
-
-	entry := &model.RecordEntry{
-		ResourceType:  model.RRouting,
-		ResourceName:  fmt.Sprintf("%s(%s)", md.Name, md.ID),
-		Namespace:     req.GetNamespace(),
-		OperationType: opt,
-		Operator:      utils.ParseOperator(ctx),
-		Detail:        detail,
-		HappenTime:    time.Now(),
-	}
-	return entry
 }
